Modularize uiSpec reducers and dedupe designer host integration
