+++ web/src/components/tabs/project/actions.tsx
@@ -13,17 +13,18 @@ import {Route} from '@/routes/_protected/projects/$projectId';
 import {ProjectStatusDialog} from '@/components/dialogs/change-project-status-dialog';
 import {useIsAuthorisedTo} from '@/hooks/auth-hooks';
 import {Action, getUserResourcesForAction} from '@faims3/data-model';
-import {useMutation, useQueryClient} from '@tanstack/react-query';
+import {useQueryClient} from '@tanstack/react-query';
 import {DesignerDialog} from '@/components/dialogs/designer-dialog';
-import type {
-  NotebookWithHistory,
-  NotebookUISpec,
-} from '@/designer/state/initial';
+import type {NotebookWithHistory} from '@/designer/state/initial';
 import {EditProjectDialog} from '@/components/dialogs/edit-project-dialog';
 import {generateTestRecordsForProject} from '@/hooks/project-hooks';
 import {Input} from '@mui/material';
 import {AddProjectToTeamDialog} from '@/components/dialogs/add-project-to-team-dialog';
 import {CreateTemplateFromProjectDialog} from '@/components/dialogs/create-tempalate-from-project-dialog';
+import {
+  toDesignerNotebookWithHistory,
+  useDesignerSaveMutation,
+} from '@/designer/integration';
 
 /**
  * ProjectActions component renders action cards for editing and closing a project.
@@ -46,42 +47,14 @@ const ProjectActions = (): JSX.Element => {
   const [generateCount, setGenerateCount] = useState('10');
   // Prepare notebook data for the Designer
   const initialNotebook = useMemo<NotebookWithHistory | undefined>(() => {
-    if (!data) return undefined;
-    return {
-      metadata: data.metadata,
-      'ui-specification': {
-        present: data['ui-specification'] as unknown as NotebookUISpec,
-        past: [],
-        future: [],
-      },
-    };
+    return toDesignerNotebookWithHistory(data);
   }, [data]);
 
-  // save updated notebook JSON
-  const saveFile = useMutation<unknown, Error, File>({
-    mutationFn: async file => {
-      const jsonText = await file.text();
-      const res = await fetch(
-        `${import.meta.env.VITE_API_URL}/api/notebooks/${projectId}`,
-        {
-          method: 'PUT',
-          headers: {
-            'Content-Type': 'application/json',
-            Authorization: `Bearer ${user?.token}`,
-          },
-          body: jsonText,
-        }
-      );
-      if (!res.ok) {
-        const err = await res.text();
-        throw new Error(`Save failed: ${res.status} ${err}`);
-      }
-      return res.json();
-    },
-    onSuccess: updated => {
-      queryClient.setQueryData(['projects', projectId], () => updated);
-      queryClient.invalidateQueries({queryKey: ['projects', projectId]});
-    },
+  const saveProjectNotebook = useDesignerSaveMutation({
+    resourceType: 'projects',
+    apiResourceType: 'notebooks',
+    resourceId: projectId,
+    token: user?.token,
   });
 
   // need to invalidate the project query after upload
@@ -90,7 +63,7 @@ const ProjectActions = (): JSX.Element => {
   };
 
   const handleEditorClose = (file?: File) => {
-    if (file) saveFile.mutate(file);
+    if (file) saveProjectNotebook.mutate(file);
     setEditorOpen(false);
   };
