+++ web/src/components/tabs/templates/actions.tsx
@@ -8,16 +8,17 @@ import {Card} from '@/components/ui/card';
 import {useAuth} from '@/context/auth-provider';
 import {useGetTemplate} from '@/hooks/queries';
 import {Route} from '@/routes/_protected/templates/$templateId';
-import {useMutation, useQueryClient} from '@tanstack/react-query';
+import {useQueryClient} from '@tanstack/react-query';
 import {DesignerDialog} from '@/components/dialogs/designer-dialog';
-import type {
-  NotebookWithHistory,
-  NotebookUISpec,
-} from '@/designer/state/initial';
+import type {NotebookWithHistory} from '@/designer/state/initial';
 import {EditTemplateDialog} from '@/components/dialogs/edit-template';
 import {Action, getUserResourcesForAction} from '@faims3/data-model';
 import {useIsAuthorisedTo} from '@/hooks/auth-hooks';
 import {AddTemplateToTeamDialog} from '@/components/dialogs/add-template-to-team-dialog';
+import {
+  toDesignerNotebookWithHistory,
+  useDesignerSaveMutation,
+} from '@/designer/integration';
 
 /**
  * TemplateActions component renders action cards for creating a project from a template,
@@ -33,48 +34,19 @@ const TemplateActions = () => {
   const queryClient = useQueryClient();
   const [editorOpen, setEditorOpen] = useState(false);
 
-  const saveFile = useMutation<unknown, Error, File>({
-    mutationFn: async file => {
-      const jsonText = await file.text();
-      const res = await fetch(
-        `${import.meta.env.VITE_API_URL}/api/templates/${templateId}`,
-        {
-          method: 'PUT',
-          headers: {
-            'Content-Type': 'application/json',
-            Authorization: `Bearer ${user?.token}`,
-          },
-          body: jsonText,
-        }
-      );
-      if (!res.ok) {
-        const err = await res.text();
-        throw new Error(`Save failed: ${res.status} ${err}`);
-      }
-      return res.json();
-    },
-    onSuccess: updated => {
-      queryClient.setQueryData(['templates', templateId], () => updated);
-      queryClient.invalidateQueries({
-        queryKey: ['templates', templateId],
-      });
-    },
+  const saveTemplateNotebook = useDesignerSaveMutation({
+    resourceType: 'templates',
+    apiResourceType: 'templates',
+    resourceId: templateId,
+    token: user?.token,
   });
 
   const initialNotebook = useMemo<NotebookWithHistory | undefined>(() => {
-    if (!data) return undefined;
-    return {
-      metadata: data.metadata,
-      'ui-specification': {
-        present: data['ui-specification'] as unknown as NotebookUISpec,
-        past: [],
-        future: [],
-      },
-    };
+    return toDesignerNotebookWithHistory(data);
   }, [data]);
 
   const handleEditorClose = (file?: File) => {
-    if (file) saveFile.mutate(file);
+    if (file) saveTemplateNotebook.mutate(file);
     setEditorOpen(false);
   };
