+++ web/src/context/auth-provider.tsx
@@ -15,7 +15,7 @@ export interface User {
     name: string;
     email: string;
     isVerified: boolean;
-    profiles: Map<string, boolean>;
+    profiles?: Map<string, boolean>;
   };
   token: string;
   refreshToken: string;
